+++ app/src/main/java/kr/ssu/ai_fitness/adapter/PersonAdapter.java
@@ -50,51 +50,13 @@ public class PersonAdapter extends RecyclerView.Adapter<PersonAdapter.ViewHolder
         //*****인자로받은 uid를 이용해서 서버로부터 사용자와 채팅할 수 있는 상대방 목록을 받아옴.
         //*****items.clear()랑 notifydatasetchanged()사용해야할 수도 있음
 
-        StringRequest stringRequest = new StringRequest(Request.Method.POST, URLs.URL_CHATLIST,
-                new Response.Listener<String>() {
-                    @Override
-                    public void onResponse(String response) {
-                        //서버에서 요청을 받았을 때 수행되는 부분
 
-                        try {
-                            //response를 json object로 변환함.
-                            JSONArray obj = new JSONArray(response);
-                            //JSONObject userJson = obj.getJSONObject(0);
+    }
 
-                            temp.add(obj.getInt(0));
-                            temp.add(obj.getInt(1));
+    public void swapDataSet(ArrayList<Person> newData) {
+        items = newData;
 
-                            Log.d("xxxxxxxxx", obj.toString());
-
-
-                        } catch (JSONException e) {
-                            e.printStackTrace();
-                        }
-
-                        notifyDataSetChanged();
-                    }
-                },
-                new Response.ErrorListener() {
-                    @Override
-                    public void onErrorResponse(VolleyError error) {
-                        Toast.makeText(context, "error", Toast.LENGTH_SHORT).show();
-                    }
-                }) {
-            @Override
-            protected Map<String, String> getParams() throws AuthFailureError {
-                //서버가 요청하는 파라미터를 담는 부분
-                Map<String, String> params = new HashMap<>();
-                params.put("uid", String.valueOf(uid));
-                return params;
-            }
-        };
-
-        //아래 큐에 add 할 때 Volley라고 하는 게 내부에서 캐싱을 해준다, 즉, 한번 보내고 받은 응답결과가 있으면
-        //그 다음에 보냈을 떄 이전 게 있으면 그냥 이전거를 보여줄수도  있다.
-        //따라서 이렇게 하지말고 매번 받은 결과를 그대로 보여주기 위해 다음과같이 setShouldCache를 false로한다.
-        //결과적으로 이전 결과가 있어도 새로 요청한 응답을 보여줌
-        stringRequest.setShouldCache(false);
-        VolleySingleton.getInstance(context).addToRequestQueue(stringRequest);
+        notifyDataSetChanged();
     }
 
     public void setOnItemClickListner(OnPersonItemClickListener listner) {
